requirements.txt
apache-airflow==2.11.0
apache-airflow-providers-http
apache-airflow-providers-sqlite
jupyterlab
jupyterhub
jupyter-server-proxy
jupyter-vscode-proxy
